Create base and add logic to functions
Comments added to label functions. Fixed problem with the new grid
failing to replace the old one. Another look at the rubric for this
project states the container should NOT be changing with the amount of
pixels that fill the container. Grid change button logic must be changed
to accommodate the guidelines of the project.